Allow user on trip to edit/delete expenses on trip that they havent created

+++ src/routes/expense/DeleteExpense.ts
@@ -1,11 +1,14 @@
 import { FastifyInstance } from 'fastify';
 import ExpenseRepository from '../../repository/ExpenseRepository';
+import type TripRepository from '../../repository/TripRepository';
 
 class DeleteExpenseRoute {
   expenseRepository: ExpenseRepository;
+  tripRepository: TripRepository;
 
-  constructor({ expenseRepository }: ContainerCradle) {
+  constructor({ expenseRepository, tripRepository }: ContainerCradle) {
     this.expenseRepository = expenseRepository;
+    this.tripRepository = tripRepository;
   }
 
   configure(server: FastifyInstance) {
@@ -19,10 +22,16 @@ class DeleteExpenseRoute {
         const { expenseId } = req.params;
         const userId = req.requestContext.get('userId')!;
 
-        const [existingExpense] = await this.expenseRepository.findExpensesForTrip({ expenseIds: [expenseId], userId });
+        const [existingExpense] = await this.expenseRepository.findExpensesForTrip({ expenseIds: [expenseId] });
 
         if (!existingExpense) {
-          return reply.code(404).send({ error: 'Trip or expense not found' });
+          return reply.code(404).send({ error: 'Not found' });
+        }
+
+        const [trip] = await this.tripRepository.getTrips({ userId, tripIds: [existingExpense.tripId] });
+
+        if (!trip) {
+          return reply.code(400).send({ error: 'Trip not found' });
         }
 
         await this.expenseRepository.deleteExpense(expenseId);
